// src/generated/get-ancillaries.ts
export interface components {
    schemas: {
// ...
        ServiceList: {
            /** @description Contains SSR and ancillary information. */
            ancillaries?: components["schemas"]["ServiceListResponse"];
            /** @description Lists the errors that occurred during request processing. */
            errors?: components["schemas"]["Error"][];
            /** @description Lists the warnings that occurred during request processing. */
            warnings?: components["schemas"]["Warning"][];
        };
// ...
        OfferItem: {
            /** @description The unique ID of an offer item. */
            offerItemId?: components["schemas"]["Identifier"];
            /** @description Reference to service defined in the same message. */
            serviceDefinitionRef: components["schemas"]["Identifier"];
            /** @description Reference to price defined in the same message. */
            priceDefinitionRef?: components["schemas"]["Identifier"];
            /** @description Lists all the segments associated with the offer item by ID. */
            segmentRefs?: components["schemas"]["Identifier"][];
            /** @description Lists all the passengers associated with the offer by ID. */
            passengerRefs: components["schemas"]["Identifier"][];
        };
// ...
        Segment: {
            /** @description The identifier of the segment within the context of one message. */
            id: components["schemas"]["Identifier"];
            /** @description The IATA two-character code of the airline that provides the passenger's booking or reservation. */
            bookingAirlineCode: components["schemas"]["AirlineCode"];
            /** @description The flight number of the airline that provides the passenger's booking or reservation. */
            bookingFlightNumber: components["schemas"]["CarrierFlightNumber"];
            /** @description The three-letter IATA code of the departure airport. */
            departureAirportCode: components["schemas"]["IataCityCode"];
            /** @description The three-letter IATA code of the arrival airport. */
            arrivalAirportCode?: components["schemas"]["IataCityCode"];
            /** @description The scheduled departure date in `YYYY-MM-DD` format in the airport's time zone. */
            departureDate: components["schemas"]["Date"];
            /** @description The scheduled departure time in `HH:mm:ss` format. */
            departureTime?: components["schemas"]["Time"];
            /** @description The scheduled arrival date in `YYYY-MM-DD` format in the airport's time zone. */
            arrivalDate?: components["schemas"]["Date"];
            /** @description The scheduled arrival time in `HH:mm:ss` format. */
            arrivalTime?: components["schemas"]["Time"];
            /** @description The two-letter IATA code or three-letter ICAO code of the airline that operates the flight and transports the passengers in a codeshare scenario. */
            operatingAirlineCode?: components["schemas"]["AirlineCode"];
            /** @description The flight number of the airline that operates the flight and transports the passengers in a codeshare scenario. */
            operatingFlightNumber?: components["schemas"]["CarrierFlightNumber"];
            /** @description The code assigned to the booking class; this is sometimes referred to as the reservation booking designator (RBD). */
            bookingClassCode?: components["schemas"]["BookingClassCode"];
            /** @description The code assigned to the booking class; this is sometimes referred to as the reservation booking designator (RBD). */
            operatingBookingClassCode?: components["schemas"]["BookingClassCode"];
            /** @description The code of the cabin. */
            cabinCode?: components["schemas"]["CabinTypeCode"];
            /**
             * @description Specifies the reservation status as listed on the PNR.
             * @example HK
             */
            reservationStatus?: string;
            /**
             * @description If true, there will be a change of aircraft for a continuing flight in this segment.
             * @example false
             */
            isChangeOfGauge?: boolean;
            /** @description Lists all the legs of the flight. This includes the departure, arrival, and aircraft ID. */
            flightLegs?: components["schemas"]["FlightLeg"][];
        };
        /** @description Contains the departure, arrival, and aircraft ID used for a leg of the segment. */
        FlightLeg: {
            /** @description The ID of the aircraft used for one of the legs from the segment. */
            legId: components["schemas"]["Identifier"];
            /** @description The three-letter IATA code of the departure airport. */
            departureAirportCode: components["schemas"]["IataCityCode"];
            /** @description The three-letter IATA code of the arrival airport. */
            arrivalAirportCode: components["schemas"]["IataCityCode"];
            /**
             * @description Specifies the three-letter IATA aircraft type designator.
             *     Refer to the [Aircraft Equipment Lookup](https://developer.sabre.com/docs/rest_apis/utility/aircraft_equipment_lookup/reference-documentation#/default/aircraftEquipmentLookup) API to obtain this information.
             * @example 7S8
             */
            equipmentCode?: string;
        };
        /** @description Contains passenger information, including ID, passenger type, and naming data. */
        Passenger: {
            /** @description Uniquely identifies a passenger in the Sabre system within the context of one message. */
            passengerId: components["schemas"]["Identifier"];
            /** @description Three-letter value that applies to the characteristic of the passenger. These typically drive pricing values. */
            passengerTypeCode: components["schemas"]["PassengerTypeCode"];
            /** @description The title used before the passenger's surname. */
            title?: components["schemas"]["Title"];
            /** @description The passenger's first name. */
            givenName?: components["schemas"]["GivenName"];
            /** @description The passenger's middle name or initial. */
            middleName?: components["schemas"]["MiddleName"];
            /** @description The passenger's family name or last name. Excludes any potential suffix. */
            surname?: components["schemas"]["Surname"];
            /** @description The passenger's suffix. */
            suffixName?: components["schemas"]["SuffixName"];
        };
        /** @description Contains the definition of the price the customer will pay. */
        PriceDefinition: {
            /** @description Uniquely identifies a `PriceDefinition` within the context of one message. */
            id: components["schemas"]["Identifier"];
            /** @description Contains the final price the customer will pay after the price modification system has determined if any markup or discount applies to the offer. */
            serviceFee?: components["schemas"]["ServiceFee"];
            /** @description Contains the price before modification. */
            baseFee?: components["schemas"]["ServiceFee"];
        };
// ...
        ServiceDefinition: {
            /** @description Uniquely identifies an ancillary definition within the context of one message. */
            id: components["schemas"]["Identifier"];
            /** @description The ancillary code, also known as an SSR code. This code is either an industry-defined or carrier-defined code identifying a specific type of service. */
            serviceCode: components["schemas"]["ServiceTypeCode"];
            /** @description The IATA two-letter or ICAO three-letter airline code of the ancillary's owning carrier. */
            airlineCode?: components["schemas"]["AirlineCode"];
            /**
             * @description Specifies the commercial name associated with the service code.
             * @example SECOND BAG UPTO50LB 23KG
             */
            commercialName?: string;
            /** @description The ATPCO group code. */
            groupCode?: components["schemas"]["GroupCodeType"];
            /** @description Specifies how and if the sub code must be booked. */
            additionalInputRequirements?: components["schemas"]["AdditionalInputRequirements"];
            /**
             * @description Specifies the type of baggage.
             * @example GOLF
             */
            ancillaryBagDescriptionCode?: string;
            /** @description The NDC airline code that is responsible for this offer. */
            ownerCode?: components["schemas"]["AirlineCode"];
            /**
             * @description Baggage weight up to limit in kilograms.
             * @example 32
             */
            upToWeightLimitInKilograms?: number;
            /**
             * @description Baggage weight up to limit in Pounds.
             * @example 70
             */
            upToWeightLimitInPounds?: number;
            /** @description Specifies the reason for issuing an EMD for the service associated to the Sub Code. */
            reasonForIssuance?: components["schemas"]["ReasonForIssuanceEnum"];
            /** @description The ATPCO subgroup. */
            subGroup?: components["schemas"]["SubGroupType"];
            /**
             * @description Maximum quantity of this service allowed by the airline.
             * @example 10
             */
            maximumQuantity?: number;
            /** @description Specifies how and if the sub code must be booked. */
            bookingMethod?: components["schemas"]["BookingMethodEnum"];
            /**
             * @description Specifies the Interline settlement method code.
             * @example DS, EA
             */
            settlementMethodCode?: string | null;
            /** @description Specifies the instructions for cabin service upgrade. */
            cabinUpgrade?: components["schemas"]["ServiceCabinUpgrade"];
            /** @description Lists descriptions of the service. */
            descriptionFreeText?: components["schemas"]["DescriptionFreeText"][];
        };
// ...
        ServiceCabinUpgrade: {
            /**
             * @description Specifies the service upgrade method code (PADIS 9750).
             * @example A
             */
            methodCode?: string | null;
            /**
             * @description Specifies the new cabin type code related to the upgrade method.
             * @example C
             */
            reservationBookingDesignator?: string;
        } | null;
        /** @description Contains the required additional input that must be provided while booking an offer. */
        AdditionalInputRequirements: {
            /**
             * @description Text-based information about the SSR entry.
             * @example %FREETEXT%
             */
            pattern: string;
            /** @description Lists the variables that may have been used in `pattern`. */
            variables?: components["schemas"]["AdditionalInputPatternVariable"][];
        };
        /** @description Contains the definition of the variable used in the additional input pattern. */
        AdditionalInputPatternVariable: {
            /**
             * @description Specifies the variable identifier as used in pattern, without wrapping the "%" characters.
             * @example FREETEXT
             */
            key: string;
            /**
             * @description Specifies the expression pattern of the variable.
             * @example [A-Z0-9\-/ ]*
             */
            value: string;
            /**
             * @description Human-readable explanation of the meaning of this SSR parameter, e.g. "Pick Up Address", "Phone Contact".  Required if the SSR parameter represents additional input mandatory from the user to book the ancillary or SSR.
             * @example Chauffeur Drive To Airport
             */
            description?: string;
        };
// ...
        DescriptionFreeText: {
            /**
             * @description The identifier of a description item.
             * @example 0
             */
            id: string;
            /**
             * @description The text of a description item.
             * @example FREE TEXT
             */
            text: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}